jack/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
decorator==5.0.7
discord==1.0.1
discord.py==1.7.1
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==2.10
multidict==5.1.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.17.0
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
typing-extensions==3.7.4.3
urllib3==1.26.4
validators==0.18.2
yarl==1.6.3
youtube-dl==2021.4.7
youtube-search-python==1.4.3
